Refactored EulerTools
Seperated the methods that delt with arrays and lists into a new classs called CollectionTools. I moved the permutation methods to the new class because they treats the String like the array of characters that underlies it, as opposed to isPalindrome(String), which feels to me like it's using the String as a higher level concept of words. Also, palindromes is a more toy property that EulerTools is supposed to contain